requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.5.0
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-vision==3.10.1
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.16.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.11.4
pydantic_core==2.33.2
pytesseract==0.3.13
python-bidi==0.6.6
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
scikit-image==0.25.2
scipy==1.15.3
setuptools==80.8.0
shapely==2.1.1
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tifffile==2025.5.10
torch==2.7.0
torchvision==0.22.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
